pyproject.toml
[project]
name = "rag-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "faiss-cpu>=1.12.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-openai>=0.3.32",
    "pandas>=2.3.2",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
]
